Move the tutorial files to subfolders
The tutorial files are moved into subfolders to better encapsulate each
tutorial. This also requires adapting the globbing pattern in the
`build_examples` tool to go into subfolders.

This change also removes the existing, monolithic tutorial/readme.md and
moves the content into a the individual tutorial files.

Finally the contents of the tutorial files is extended to better explain
what is going on in each example.